binary4.04

+++ pr_int.c
@@ -10,7 +10,7 @@
 int pr_int(va_list args, t_flag *fl)
 {
 	int i, num, y;
-	int x;
+	long int x;
 	char s[40] = {0};
 
 	x = va_arg(args, long int);
